Try to fix tensorflow issues on windows. Update doc version number

+++ umap/parametric_umap.py
@@ -826,8 +826,8 @@ def construct_edge_dataset(
 
     # shuffle edges
     shuffle_mask = np.random.permutation(range(len(edges_to_exp)))
-    edges_to_exp = edges_to_exp[shuffle_mask]
-    edges_from_exp = edges_from_exp[shuffle_mask]
+    edges_to_exp = edges_to_exp[shuffle_mask].astype(np.int64)
+    edges_from_exp = edges_from_exp[shuffle_mask].astype(np.int64)
 
     # create edge iterator
     if parametric_embedding:
